+++ src/responses/child_device_list_result/s200b_result.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::api::{Authenticated, HubHandler};
+use crate::api::HubHandler;
 use crate::error::Error;
 use crate::requests::{EmptyParams, GetTriggerLogsParams, TapoParams, TapoRequest};
 use crate::responses::{
@@ -80,10 +80,7 @@ pub enum S200BLog {
 impl S200BResult {
     /// Gets *device info* as [`crate::responses::S200BResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API.
-    pub async fn get_device_info(
-        &self,
-        handler: &HubHandler<Authenticated>,
-    ) -> Result<S200BResult, Error> {
+    pub async fn get_device_info(&self, handler: &HubHandler) -> Result<S200BResult, Error> {
         let request = TapoRequest::GetDeviceInfo(TapoParams::new(EmptyParams));
 
         handler.control_child(self.device_id.clone(), request).await
@@ -99,7 +96,7 @@ impl S200BResult {
     /// Use a `start_id` of `0` to get the most recent X logs, where X is capped by `page_size`.
     pub async fn get_trigger_logs(
         &self,
-        handler: &HubHandler<Authenticated>,
+        handler: &HubHandler,
         page_size: u64,
         start_id: u64,
     ) -> Result<TriggerLogsResult<S200BLog>, Error> {