+++ src/responses/child_device_list_result/t100_result.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::api::{Authenticated, HubHandler};
+use crate::api::HubHandler;
 use crate::error::Error;
 use crate::requests::{EmptyParams, GetTriggerLogsParams, TapoParams, TapoRequest};
 use crate::responses::{
@@ -61,10 +61,7 @@ pub enum T100Log {
 impl T100Result {
     /// Gets *device info* as [`crate::responses::T100Result`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API.
-    pub async fn get_device_info(
-        &self,
-        handler: &HubHandler<Authenticated>,
-    ) -> Result<T100Result, Error> {
+    pub async fn get_device_info(&self, handler: &HubHandler) -> Result<T100Result, Error> {
         let request = TapoRequest::GetDeviceInfo(TapoParams::new(EmptyParams));
 
         handler.control_child(self.device_id.clone(), request).await
@@ -80,7 +77,7 @@ impl T100Result {
     /// Use a `start_id` of `0` to get the most recent X logs, where X is capped by `page_size`.
     pub async fn get_trigger_logs(
         &self,
-        handler: &HubHandler<Authenticated>,
+        handler: &HubHandler,
         page_size: u64,
         start_id: u64,
     ) -> Result<TriggerLogsResult<T100Log>, Error> {